Add webpack fallback config for `fs` and `path`
The spirit of the fallback configuration is to do the expected thing when encountering a native nodejs require. Having this fallback configuration
allows for more uses of `import` within an extension without having additional custom webpack config in the config.ts file.

This case specifically allows for importing emscripten generated javascript files which import fs or path but have ways of using them that don't invoke those codepaths.