Add webpack fallback config for `fs` and `path` (#70)
The spirit of our fallback configuration is to do the expected thing when encountering a native nodejs require. Having the fallback configuration in this change allows for more uses of import within an extension without having additional custom webpack config in the config.ts file.

This case specifically allows for importing emscripten generated javascript files which import fs or path but have ways of using them that don't invoke those codepaths.