Update examples and template to use object subscribe syntax
The string array form of subscribe has the legacy behavior of subscribing with preloading. For most custom panels this is not necessary and results in requesting more data. This change updates the template and examples to use the new subscribe api which gives the user the option of which topics are preloaded.