Update examples to unmount react component on panel teardown (#93)
The `initPanel` function can return a cleanup function that is run when
the panel is torn down (aka removed). A common use of this function is
to cleanup event handlers or unmount components to free up resources. We
updated Studio panels to return a cleanup function.

This change updates the examples to return a cleanup function so
extension authors benefit from having a way to cleanup resources and
unmount components.