Update local-install to create `$HOME/.foxglove-studio` folder if snap is not present (#99)
In #91 we updated the install logic to install correctly for snap users.
In that logic we check for the presence of specific directories before
installing (whereas before we would always make the directory). Since
then we've learned that debian package users don't actually have a
`$HOME/.foxglove-studio/extensions` directory prior to actually
installing an extension. This meant that the `local-install` script
would fail to install a developer's extension locally unless they had
installed one via the marketplace or made the extensions dir manually.

This change updates the `install` command to pro-actively create the
`$HOME/.foxglove-studio/extensions` directory if it does not exist. To
avoid creating this directory for _snap_ users, we first check for the
presence of the _snap/foxglove-studio/current_ directory which tells us
if a snap is installed for foxglove-studio. If the snap is used then we
only install in the snap directory.

We'll see how this holds up to other create ways folks expected the
local install command to work.

Fixes: #98