Improve e2e CI workflow (#161)
Improve the e2e tests by using a packaged build of the
`create-foxglove-extension` package. This packaged build tarball is used
with the `npx` command to test the `npx init` functionality. This same
tarball is installed into the `sample` extension to confirm the
`package` script specified in the sample package.json works.