Hackathon feedback (#168)
### Changelog
None.
### Docs
None.
<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

Tried using this for hackathon, came across a few issues:

1. There are several references to Studio in the generated comments and
in linked docs, I updated those.
1. The only other reference to Studio that I thought seemed clearly
out-of-date was in the publish code which sets the README and changelog
links. This seemed trivially broken, since the `readme` and `changelog`
locals are not read after they're set. I'm not sure why a linter didn't
catch this? Either way - I have not tested this functionality out, so if
it seems potentially incorrect, I can take that commit out.
<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->