rust data loader example (#171)
### Changelog

Adds rust data loader example.

### Docs

None

### Description

This PR adds a rust data loader example with the example file from the
`foxglove_data_loader` in
[foxglove-sdk/#460](https://github.com/foxglove/foxglove-sdk/pull/460)
plus the typescript build to generate the extension.

The example in the sdk repo is rust-only to keep that repo simpler and
so that users can understand only that rust code in isolation, but the
example here in create-foxglove-extension is more complete and something
we can point people to for everything they will need to make an
extension of their own.

Replaces
[#169](https://github.com/foxglove/create-foxglove-extension/pull/169).

Fixes: FG-11789

Depends on https://github.com/foxglove/foxglove-sdk/pull/460