remove prior extensions from a list by id instead of directory name (#173)
### Changelog


<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

There can be multiple versions of an extension installed under different
capitalizations:

```
$ ls ~/.foxglove-studio/extensions
 foxglove.foxglove-extension-demo-1.1.0   foxglove.Foxglove-EXTENSION-Demo-1.1.0
 foxglove.Foxglove-Extension-Demo-1.1.0  'Robert Holt.staticSceneModelConverter-0.0.3'
```

When there are multiple extension directories like this, desktop will
load the first matching extension id by the default lexicographic
sorting of `readdir()`, which happens to sort capital letters before
lower case:

```
$ ls /tmp/x
a  A  b  B
$ node
> fs.readdirSync('.')
[ 'A', 'B', 'a', 'b' ]
```

The capitalizations were from the previous behavior, so these extensions
will tend to be older than newer extensions, causing users to notice
that extensions weren't updating properly.

Before, the old capitalizations would linger after doing a
`foxglove-extension install`. With this patch, after `foxglove-extension
install`, the other variants with the same ID are removed as expected:

```
$ ls ~/.foxglove-studio/extensions
foxglove.foxglove-extension-demo-1.1.0  'Robert Holt.staticSceneModelConverter-0.0.3'
```

The current code uses a directory name calculated from the extension
package.json name and publisher, but this was handled inconsistently
between web and desktop versions of foxglove until recently, so there
may be different directories for the same id in a given user extension
directory. And `foxglove-extension install` also does not normalize the
directory path the same as the upstream changes to desktop that always
lower case the name component of the extension directory path.

This patch includes the `listExtensions()` routine from desktop and
removes all extensions from the list that match the ID that the user is
trying to install before copying extension files into the user extension
directory. The previous behavior would only remove a previous
extension's files when the directory name exactly matched the new
calculated directory path.

Fixes: FG-12249

---------

Co-authored-by: Hans-Joachim Krauch <achim.krauch@gmail.com>