Add csv data loader example (#179)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Adds another example data loader for loading from a CSV. It uses an
index to keep track of where to start reading from. It logs each column
as a different topic and uses a "timestamp_nanos" column for the log
time.