Add wasm rule to shared config (#181)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

If any extension is using WASM (like the data loader extensions) they'll
need to inline it as base64. This removes the need for people to do this
manually by adding it to shared config.

Fixes FIRE-82.