Add resourceQuery to allow for importing source files as raw strings (#183)
This change updates our webpack configuration to support importing
typescript source files as a string. This makes it possible to write a
typescript source file that you can then import as a string for use as
an argument to extension apis.

```typescript
import foo from "foo.ts?raw";

extensionApiToRegisterFoo(foo);
```